requirements.txt
python-binance
happybase
kafka-python
pytest
tqdm
pyspark
pandas
dash
fastapi
httptools==0.1.*
uvloop
uvicorn